requirements.txt
torch
torchvision
safetensors
pillow
pandas
pydantic
pydantic-settings
tqdm
streamlit
scikit-learn
matplotlib
